alpine-consul-python-example/root/app/requirements.txt
python-consul
pyramid